@@ -8,10 +8,8 @@
     "@fullcalendar/interaction": "^6.1.8",
     "@fullcalendar/react": "^6.1.8",
     "@fullcalendar/timegrid": "^6.1.8",
-    "@types/dotenv": "^8.2.0",
-    "@types/moment": "^2.13.0",
+    "axios": "^1.4.0",
     "bootstrap": "~4.6.1",
-    "connected-react-router": "^6.9.3",
     "dotenv": "^16.3.1",
     "eonasdan-bootstrap-datetimepicker": "^4.17.49",
     "history": "4.10.1",
@@ -20,39 +18,40 @@
     "moment": "^2.29.4",
     "moment-timezone": "^0.5.43",
     "popper.js": "^1.16.1",
-    "react": "~16.14.0",
+    "react": "^18.2.0",
     "react-big-calendar": "^1.8.1",
-    "react-bootstrap": "^2.7.4",
+    "react-bootstrap": "^1.6.6",
     "react-datepicker": "^4.11.0",
     "react-datetime-picker": "^5.0.3",
-    "react-dom": "~16.14.0",
-    "react-redux": "~7.2.6",
-    "react-router": "~5.2.1",
-    "react-router-dom": "~5.3.0",
+    "react-dom": "^18.2.0",
+    "react-redux": "^8.0.4",
+    "react-router": "^6.4.2",
+    "react-router-dom": "^6.4.2",
+    "react-scripts": "^5.0.1",
     "react-select": "^5.7.3",
     "react-toastify": "^9.1.3",
     "reactstrap": "^9.2.0",
     "redux": "4.1.2",
+    "redux-first-history": "^5.1.1",
     "redux-thunk": "2.4.1",
     "styled-components": "^6.0.3"
   },
   "devDependencies": {
-    "@types/history": "4.7.10",
-    "@types/jest": "24.0.19",
+    "@types/history": "^4.7.10",
+    "@types/jest": "^24.0.19",
     "@types/node": "^12.20.55",
-    "@types/react": "18.0.35",
+    "@types/react": "18.2.0",
     "@types/react-big-calendar": "^1.6.4",
     "@types/react-datepicker": "^4.10.0",
-    "@types/react-dom": "~16.9.14",
-    "@types/react-redux": "~7.1.22",
-    "@types/react-router": "~5.1.17",
-    "@types/react-router-dom": "~5.3.0",
-    "@types/reactstrap": "~8.7.2",
+    "@types/react-dom": "^18.2.0",
+    "@types/react-redux": "^7.1.25",
+    "@types/react-router": "^5.1.20",
+    "@types/react-router-dom": "^5.3.3",
     "cross-env": "6.0.3",
     "eslint-plugin-jsx-a11y": "^6.4.1",
     "nan": "^2.15.0",
-    "react-scripts": "^5.0.0",
-    "typescript": "3.6.4"
+    "react-scripts": "^5.0.1",
+    "typescript": "^4.8.2"
   },
   "scripts": {
     "start": "react-scripts start",
