@@ -1,27 +1,43 @@
-import 'bootstrap/dist/css/bootstrap.css';
+//import 'bootstrap/dist/css/bootstrap.css';
+//import * as React from 'react';
+//import * as ReactDOM from 'react-dom';
+//import { Provider } from 'react-redux';
 
+//import { ConnectedRouter } from 'connected-react-router';
+//import { createBrowserHistory } from 'history';
+//import configureStore from './store/configureStore';
+//import App from './App';
+//import registerServiceWorker from './registerServiceWorker';
+
+//// Create browser history to use in the Redux store
+//const baseUrl = document.getElementsByTagName('base')[0].getAttribute('href') as string;
+//const history = createBrowserHistory({ basename: baseUrl });
+
+//// Get the application-wide store instance, prepopulating with state from the server where available.
+//const store = configureStore(history);
+
+//ReactDOM.render(
+//    <Provider store={store}>
+//        {/*<ConnectedRouter history={history}>*/}
+//            <App/>
+//        {/*</ConnectedRouter>*/}
+//    </Provider>,
+//    document.getElementById('root'));
+
+//registerServiceWorker();
+import 'bootstrap/dist/css/bootstrap.css';
 import * as React from 'react';
-import * as ReactDOM from 'react-dom';
-import { Provider } from 'react-redux';
-import { ConnectedRouter } from 'connected-react-router';
-import { createBrowserHistory } from 'history';
-import configureStore from './store/configureStore';
+import * as reactdomclient from 'react-dom/client'; // Import createRoot from react-dom
+import { BrowserRouter as Router } from 'react-router-dom';
 import App from './App';
 import registerServiceWorker from './registerServiceWorker';
 
-// Create browser history to use in the Redux store
-const baseUrl = document.getElementsByTagName('base')[0].getAttribute('href') as string;
-const history = createBrowserHistory({ basename: baseUrl });
+const rootElement = document.getElementById('root') as Element; // Use type assertion
 
-// Get the application-wide store instance, prepopulating with state from the server where available.
-const store = configureStore(history);
+// Use createRoot instead of ReactDOM.render
+reactdomclient.createRoot(rootElement).render(
 
-ReactDOM.render(
-    <Provider store={store}>
-        <ConnectedRouter history={history}>
-            <App/>
-        </ConnectedRouter>
-    </Provider>,
-    document.getElementById('root'));
+    <App />
+);
 
 registerServiceWorker();
